Add Voice settings page with mic and speaker device selection
Split device selection and always-listen toggle into a dedicated
Voice page. Mic input device, speaker output device, and always-listen
toggle. TTS & ASR page retains model/language config.
