+++ src/components/settings/TTSSection.tsx
@@ -23,7 +23,6 @@ import AddIcon from '@mui/icons-material/Add';
 import { useTTS } from '../../hooks/useTTS';
 import { storage } from '../../utils/storage';
 import { apiClient } from '../../api/client';
-import { useMicStore } from '../../store/micStore';
 
 interface ModelEntry {
   id: string;
@@ -49,25 +48,11 @@ export const TTSSection: React.FC = () => {
   const { backends, loadBackends } = useTTS();
 
   // ASR settings
-  const [micDevices, setMicDevices] = useState<Array<{ deviceId: string; label: string }>>([]);
-  const [selectedMicId, setSelectedMicIdState] = useState(storage.getASRDeviceId() || '');
-  const [alwaysListen, setAlwaysListenState] = useState(storage.getASRAlwaysListen());
   const [asrMode, setAsrModeState] = useState(storage.getASRMode());
   const [asrLanguage, setAsrLanguageState] = useState(storage.getASRLanguage() || '');
   const [asrFixedModel, setAsrFixedModelState] = useState(storage.getASRFixedModel());
   const [modelMap, setModelMapState] = useState(storage.getASRModelMap());
   const [availableModels, setAvailableModels] = useState<ModelEntry[]>([]);
-
-  const setAlwaysListen = (v: boolean) => {
-    setAlwaysListenState(v);
-    storage.setASRAlwaysListen(v);
-    const mic = useMicStore.getState();
-    if (v && mic.status === 'idle') {
-      mic.startListening();
-    } else if (!v && mic.status !== 'idle') {
-      mic.stopListening();
-    }
-  };
 
   const setAsrMode = (v: 'fixed' | 'routing') => {
     setAsrModeState(v);
@@ -102,32 +87,10 @@ export const TTSSection: React.FC = () => {
   const setTtsEmotionAlpha = (v: number) => { setTtsEmotionAlphaState(v); storage.setTTSEmotionAlpha(v); };
   const setTtsUseEmotionText = (v: boolean) => { setTtsUseEmotionTextState(v); storage.setTTSUseEmotionText(v); };
 
-  const setSelectedMicId = (deviceId: string) => {
-    setSelectedMicIdState(deviceId);
-    if (deviceId) {
-      storage.setASRDeviceId(deviceId);
-      useMicStore.getState().selectDevice(deviceId);
-    }
-  };
-
   useEffect(() => {
     loadBackends();
     apiClient.getASRModels()
       .then((models) => setAvailableModels(models.map((m) => ({ id: m.id, name: m.name }))))
-      .catch(() => {});
-    // Load mic devices
-    navigator.mediaDevices.getUserMedia({ audio: true })
-      .then((stream) => {
-        stream.getTracks().forEach((t) => t.stop());
-        return navigator.mediaDevices.enumerateDevices();
-      })
-      .then((devices) => {
-        setMicDevices(
-          devices
-            .filter((d) => d.kind === 'audioinput')
-            .map((d) => ({ deviceId: d.deviceId, label: d.label || `Microphone ${d.deviceId.slice(0, 8)}` }))
-        );
-      })
       .catch(() => {});
   }, [loadBackends]);
 
@@ -148,43 +111,6 @@ export const TTSSection: React.FC = () => {
   return (
     <Box sx={{ maxWidth: 700 }}>
       <Typography variant="h3" sx={{ mb: 3 }}>TTS & ASR</Typography>
-
-      {/* Always Listen */}
-      <Box sx={{ mb: 3 }}>
-        <FormControlLabel
-          control={
-            <Switch
-              checked={alwaysListen}
-              onChange={(e) => setAlwaysListen(e.target.checked)}
-            />
-          }
-          label="Always listen"
-        />
-        <Typography variant="caption" color="text.secondary" sx={{ display: 'block', mt: 0.5 }}>
-          Keep microphone active to detect trigger words or push-to-talk (Ctrl+Space). Disable to save resources.
-        </Typography>
-      </Box>
-
-      {/* Microphone Selection */}
-      <Box sx={{ mb: 3 }}>
-        <FormControl fullWidth size="small">
-          <InputLabel>Microphone</InputLabel>
-          <Select
-            value={selectedMicId}
-            label="Microphone"
-            onChange={(e: SelectChangeEvent) => setSelectedMicId(e.target.value)}
-          >
-            <MenuItem value="">
-              <em>Default</em>
-            </MenuItem>
-            {micDevices.map((d) => (
-              <MenuItem key={d.deviceId} value={d.deviceId}>
-                {d.label}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-      </Box>
 
       {/* ASR Mode */}
       <Box sx={{ mb: 3 }}>
